@@ -15,47 +15,46 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^7.2.0",
-    "@angular/common": "^7.2.0",
-    "@angular/compiler": "^7.2.0",
-    "@angular/core": "^7.2.0",
-    "@angular/forms": "^7.2.0",
-    "@angular/http": "^7.2.0",
-    "@angular/platform-browser": "^7.2.0",
-    "@angular/platform-browser-dynamic": "^7.2.0",
-    "@angular/router": "^7.2.0",
-    "core-js": "^2.5.4",
-    "cytoscape": "^3.3.1",
+    "@angular/animations": "^8.2.3",
+    "@angular/common": "^8.2.3",
+    "@angular/compiler": "^8.2.3",
+    "@angular/core": "^8.2.3",
+    "@angular/forms": "^8.2.3",
+    "@angular/http": "^7.2.15",
+    "@angular/platform-browser": "^8.2.3",
+    "@angular/platform-browser-dynamic": "^8.2.3",
+    "core-js": "^3.2.1",
+    "cytoscape": "^3.9.1",
     "cytoscape-dagre": "^2.2.2",
     "dagre": "^0.8.4",
-    "rxjs": "~6.3.3",
-    "zone.js": "~0.8.27"
+    "rxjs": "~6.5.2",
+    "zone.js": "~0.10.2"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.12.1",
-    "@angular-devkit/build-ng-packagr": "^0.12.1",
-    "@angular/cli": "^7.2.1",
-    "@angular/compiler-cli": "^7.2.0",
-    "@angular/language-service": "^7.2.0",
+    "@angular-devkit/build-angular": "^0.803.0",
+    "@angular-devkit/build-ng-packagr": "^0.803.0",
+    "@angular/cli": "^8.3.0",
+    "@angular/compiler-cli": "^8.2.3",
+    "@angular/language-service": "^8.2.3",
     "@types/cytoscape": "^3.2.9",
     "@types/dagre": "^0.7.40",
-    "@types/jasmine": "~2.8.8",
+    "@types/jasmine": "~3.4.0",
     "@types/jasminewd2": "~2.0.3",
-    "@types/node": "~8.9.4",
-    "codelyzer": "~4.5.0",
-    "jasmine-core": "~2.99.1",
+    "@types/node": "~12.7.2",
+    "codelyzer": "~5.1.0",
+    "jasmine-core": "~3.4.0",
     "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~3.0.0",
-    "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "~2.0.1",
-    "karma-jasmine": "~1.1.2",
-    "karma-jasmine-html-reporter": "^0.2.2",
-    "ng-packagr": "^4.4.5",
-    "protractor": "~5.4.1",
-    "ts-node": "~7.0.0",
-    "tsickle": "^0.34.0",
+    "karma": "^4.3.0",
+    "karma-chrome-launcher": "~3.1.0",
+    "karma-coverage-istanbul-reporter": "~2.1.0",
+    "karma-jasmine": "~2.0.1",
+    "karma-jasmine-html-reporter": "^1.4.2",
+    "ng-packagr": "^5.5.0",
+    "protractor": "~5.4.0",
+    "ts-node": "~8.3.0",
+    "tsickle": "^0.37.0",
     "tslib": "^1.9.0",
-    "tslint": "~5.11.0",
-    "typescript": "3.1.6"
+    "tslint": "~5.19.0",
+    "typescript": "3.5.3"
   }
 }
